rancher-logging: Fix generated YAML for logging output fields that are referencing file paths
Several logging output fields for ca and client certificates need to be provied as a file path in the fluentd config. For this, the secret which contains the value must be referenced in the output yaml with mountFrom instead of valueFrom. See https://banzaicloud.com/docs/one-eye/logging-operator/configuration/plugins/outputs/secret/.

Addresses https://github.com/rancher/rancher/issues/31810

Signed-off-by: Bastian Hofmann <bashofmann@gmail.com>